consumer/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
annotated-types==0.5.0
    # via pydantic
anyio==3.7.1
    # via
    #   httpcore
    #   starlette
certifi==2023.7.22
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.2.0
    # via requests
click==8.1.6
    # via
    #   pact-python
    #   uvicorn
fastapi==0.100.1
    # via pact-python
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==0.16.3
    # via httpx
httpx==0.23.3
    # via pact-python
idna==3.4
    # via
    #   anyio
    #   requests
    #   rfc3986
iniconfig==2.0.0
    # via pytest
packaging==23.1
    # via pytest
pact-python==2.0.0
    # via -r requirements.in
pluggy==1.2.0
    # via pytest
psutil==5.9.5
    # via pact-python
pydantic==2.1.1
    # via fastapi
pydantic-core==2.4.0
    # via pydantic
pytest==7.4.0
    # via -r requirements.in
requests==2.31.0
    # via pact-python
rfc3986[idna2008]==1.5.0
    # via httpx
six==1.16.0
    # via pact-python
sniffio==1.3.0
    # via
    #   anyio
    #   httpcore
    #   httpx
starlette==0.27.0
    # via fastapi
typing-extensions==4.7.1
    # via
    #   fastapi
    #   pydantic
    #   pydantic-core
urllib3==2.0.4
    # via
    #   pact-python
    #   requests
uvicorn==0.23.1
    # via pact-python
